requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-core==2.20.1
SQLAlchemy==2.0.34
psycopg[binary,pool]==3.2.3
alembic==1.13.2
redis==5.0.8
celery==5.4.0
boto3==1.35.29
minio==7.2.7
prometheus-client==0.21.0
python-dateutil==2.9.0.post0
pandas==2.2.2
openpyxl==3.1.5
rapidfuzz==3.9.7
opencv-python-headless==4.10.0.84
paddleocr==2.9.1
pillow==10.4.0
python-multipart==0.0.9
orjson==3.10.7
httpx==0.27.2
typer==0.12.5
pytest==8.3.2
pytest-cov==5.0.0
ruff==0.6.8

